bump shade and unpin requirements
Addressed the issue of ending up with a broken
shade because of the pinning requirments.

If we find a breaking bug, we can start banning these versions.

Note that now shade has a gate that tests ansible compatibilty
which means is less likely to introduce problems like in the past.

The more real danger is that we downgrate it due to pinning
and we endup with a broken one because one shade own reps
is newer and incompatible with the downtraded shade. This
is what happened with keystoneauth1.

Change-Id: I30bc56889fdc34208cfdf0887b4e998f8f9abb3f