Update patch set 1
Patch Set 1:

Oleksii, the patch is good for the idea you wish to implement. But I was wondering, why not we in a new way create new params to pull the packages from the RDO trunk repo on the go. For eg. here we hardcoded master to queens, in reality, it should point to working master branch: https://trunk.rdoproject.org/centos7-master/current/

We will have to re-work this template, I can take the task but, I can work on it only after two weeks from now.

Patch-set: 1